requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.9.0
google-generativeai>=0.3.0 